[FP-55] Rename: 아이콘 파일 이동
- svgs에서 icons 폴더로 이동
